feat: Refactor fleet UI and extract Garage component + rmv test flags
Moved fleet management logic and UI from wrapper.tsx to a new components/fleet/garage.tsx file for better separation of concerns. Removed the test mode alert from app pages. The Wrapper component now handles only compliance checks and renders the Garage component if the user is compliant.
